use a crypto util file as an abstraction
