+++ lib/setup.js
@@ -1,26 +1,8 @@
 const fs = require('fs')
-const sodium = require('sodium-native')
 const Dat = require('dat-node')
 const createDir = require('./utils/createDir')
+const crypto = require('./utils/crypto')
 const json = require('./utils/json')
-
-// use libsodium to generate an id
-function uuid () {
-  const id = sodium.malloc(16)
-  sodium.randombytes_buf(id)
-  sodium.mlock(id)
-  return id
-}
-
-// use libsodium to generate a keypair
-function createKeyPair () {
-  const pubKey = sodium.malloc(sodium.crypto_box_PUBLICKEYBYTES)
-  const privKey = sodium.malloc(sodium.crypto_box_SECRETKEYBYTES)
-  sodium.crypto_box_keypair(pubKey, privKey)
-  sodium.mlock(pubKey)
-  sodium.mlock(privKey)
-  return [pubKey, privKey]
-}
 
 // Setup a new user
 module.exports = function setup ({path, name, pass}, cb) {
@@ -44,11 +26,11 @@ module.exports = function setup ({path, name, pass}, cb) {
   }
 
   // generate and write a user ID to their public dir
-  user.id = uuid()
+  user.id = crypto.uuid()
   fs.writeFileSync(path + '/public/id', user.id.toString('hex'))
 
   // generate the private and public keys
-  const [pubKey, privKey] = createKeyPair()
+  const [pubKey, privKey] = crypto.createKeyPair()
   user.pubKey = pubKey
   user.privKey = privKey
   fs.writeFileSync(path + '/public/pubkey', pubKey.toString('hex'))
@@ -74,4 +56,3 @@ module.exports = function setup ({path, name, pass}, cb) {
     dat.joinNetwork({}, err => cb(user))
   })
 }
-
